Fixes various issues of the NN, regarding the use of an adaptable bias when training with all training dataset at once
The bias on each layer should be one-dimensional, the previous behavior was to take a different bias for each of the
examples of the training set, yielding a very accurate result because it was adapting for each of the examples with
a different bias each.

The other change is the calculation of the MSE when using a very big dataset (e.g. mnist). The current behavior uses
np.linalg.norm to calculate the MSE, the previous behaviour did a dot product of the whole matrix with its traspose,
yielding a matrix in memory of O(n*n) space (mnist has 60k training samples).